chore: extract constant
