+++ packages/api/src/prividium/constants.ts
@@ -1,0 +1,2 @@
+export const EVENT_PERMISSION_RULES_FINGERPRINT =
+  "ffad968c66c9f519c2fe7b775b721e466bbf3c451194992c3c202f4b189b87c5" as const;
